backend/pyproject.toml
[project]
name = "fablescribe-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
    "deepgram-sdk>=3.9",
    "silero-vad>=5.1",
    "numpy>=1.26",
    "onnxruntime>=1.19",
    "supabase>=2.11",
    "python-jose[cryptography]>=3.3",
    "anthropic>=0.52",
    "httpx>=0.27",
    "qdrant-client>=1.12",
    "nomic>=3.3",
    "python-multipart>=0.0.18",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
